Change DesiredInstances fields to uint64
Convert `DesiredInstances` and `UnclampedDesiredInstances` in
`ScalingDecision` from `int` to `uint64` for consistency with other
integer fields in the function package (e.g., `MinInstances`,
`MaxInstances`, `CPUUsageMilli`, `MemoryUsageMiB`).

This also updates the `Recommendation` struct, client/server
serialization, key definitions, and all arithmetic operations to use
`uint64` throughout. Comparison functions now use `cmp.Compare` to
avoid underflow issues with unsigned subtraction.

+++ internal/fixture/controller.go
@@ -14,7 +14,7 @@ import (
 
 type (
 	InstanceHandler  func(ctx context.Context, fn *function.Function, excludeInstanceNames ...string) (*function.Instance, error)
-	ScaleHandler     func(ctx context.Context, fn *function.Function, desiredInstances int, reason function.ScalingReason) ([]*function.Instance, error)
+	ScaleHandler     func(ctx context.Context, fn *function.Function, desiredInstances uint64, reason function.ScalingReason) ([]*function.Instance, error)
 	HeartbeatHandler func(ctx context.Context, routerIP string, heartbeats []*function.Heartbeat, forwardedFor ...string) error
 )
 
@@ -79,7 +79,7 @@ func (f *MockControllerClient) Instance(ctx context.Context, fn *function.Functi
 }
 
 // Scale implements controller.Client.
-func (f *MockControllerClient) Scale(ctx context.Context, fn *function.Function, desiredInstances int, reason function.ScalingReason) ([]*function.Instance, error) {
+func (f *MockControllerClient) Scale(ctx context.Context, fn *function.Function, desiredInstances uint64, reason function.ScalingReason) ([]*function.Instance, error) {
 	if f.scaleHandler == nil {
 		f.t.Fatalf("mcc.Scale was called but not mocked")
 	}
